refactor(#1701): split the stream-boundary read-metric tests into a sibling file
The lite gate's file-size ratchet caught joined_scan_stream.rs crossing the
~800-line source target (627 -> 868) once the Drop accounting and its two tests
landed. The tests move to joined_scan_stream_read_metric_tests.rs, included via
#[path] exactly like the sibling *_tests.rs modules in this directory, leaving
the source file at 722 lines. No test content changed; both cases still pass
from the new location.
